+++ matrix_files/boolean/manipulation/mult_row.cpp
@@ -2,10 +2,10 @@
 
 void Matrix<bool>::multr(unsigned row, bool multiplier) {
   row--;
-  if (row < _rows_ && !multiplier) {
+  if (static_cast<int64_t>(row) < _rows_ && !multiplier) {
     uint64_t mask = Matrix::rvm[row & 7];
     row >>= 3;
-    for (int i = _lenc_ * row; i < _lenc_ * (row + 1); i++) {
+    for (uint64_t i = _lenc_ * row; i < _lenc_ * (row + 1); i++) {
       _data_[i] &= ~mask;
     }
   }
